+++ src/icpp-gadget.cpp
@@ -316,7 +316,7 @@ __ICPP_EXPORT__ extern "C" int icpp_gadget(int argc, char **argv) {
   cl::ParseCommandLineOptions(
       argc, argv,
       std::format(
-          "ICPP, Interpreting C++, running C++ in anywhere like a script.\n"
+          "ICPP, Interpreting C++, running C++ anywhere like a script.\n"
           "  Remote icpp-gadget server built with ICPP {}",
           icpp::version_string()));