+++ src/icpp-pad.cpp
@@ -415,7 +415,7 @@ int main(int argc, char **argv) {
   cl::ParseCommandLineOptions(
       argc, argv,
       std::format(
-          "ICPP, Interpreting C++, running C++ in anywhere like a script.\n"
+          "ICPP, Interpreting C++, running C++ anywhere like a script.\n"
           "  IObject Launch Pad Tool built with ICPP {}",
           icpp::version_string()));